# docs/swagger.yaml
basePath: /
definitions:
...
  dto.PaginatedUserResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RoleResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      id:
        example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      name:
        example: Admin
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      name:
        example: John Doe Updated
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  dto.UpdateUserRolesRequest:
    properties:
      role_ids:
        example:
        - a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        - f9e8d7c6-b5a4-3210-fedc-ba0987654321
        items:
          type: string
        minItems: 1
        type: array
    required:
    - role_ids
    type: object
  dto.UserResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
  dto.UserWithRoleResponse:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: a1b2c3d4-e5f6-7a8b-9c0d-1e2f3a4b5c6d
        type: string
      name:
        example: John Doe
        type: string
      roles:
        items:
          $ref: '#/definitions/dto.RoleResponse'
        type: array
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
  response.ApiResponse:
    properties:
      data: {}
      error:
        example: Error description
        type: string
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: Production-ready API boilerplate built with Gin, GORM, and Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gin Boilerplate API
  version: "1.0"
paths:
...
  /api/v1/users:
    get:
      description: Get paginated list of users with optional filtering and sorting.
        Requires 'view_user' permission.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Sorting criteria (e.g. name:asc, created_at:desc or -created_at)
        in: query
        name: sort
        type: string
      - description: Filter by user name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by user email (partial match)
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginatedUserResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires view_user permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid filter or sorting parameter
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
  /api/v1/users/{id}:
    get:
      description: Retrieve detailed user profile including assigned roles by UUID.
        Requires 'view_user' permission.
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserWithRoleResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires view_user permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid UUID format
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - Users
  /api/v1/users/{id}/roles:
    delete:
      consumes:
      - application/json
      description: Remove one or more assigned roles from a user by role UUIDs. Requires
        'remove_user_role' permission.
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Role UUIDs to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Roles removed from user successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires remove_user_role permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid UUID format or request payload
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Remove roles from user
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Assign one or more roles to a user by role UUIDs. Requires 'add_user_role'
        permission.
      parameters:
      - description: User UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Role UUIDs to assign
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Roles added to user successfully
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "403":
          description: Forbidden - Requires add_user_role permission
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid UUID format or request payload
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Add roles to user
      tags:
      - Users
  /api/v1/users/profile:
    get:
      description: Get current authenticated user profile details
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update current authenticated user profile details (e.g. name)
      parameters:
      - description: User profile update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized - Missing or invalid token
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "404":
          description: Not Found - User not found
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "422":
          description: Unprocessable Entity - Invalid request payload
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token (e.g. "Bearer eyJhbGci...").
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"